requirements.txt
streamlit>=1.28.0
requests>=2.31.0
folium>=0.14.0
streamlit-folium>=0.15.0
python-dotenv>=1.0.0
plotly>=5.0.0

